Added favicon(s). Removed cheepr routes & added "home" route to new "index.js" routes. Removed cheepr layout. Added bootstrap mat. des. layout to login and home. Changed to fixed navbar.
