--- views/cheep.handlebars
@@ -1,4 +1,0 @@
-<div class="cheep" user="{{newcheep.username}}" id="{{newcheep._id}}" uid="{{newcheep.userid}}">
-  <b>{{newcheep.username}} said... </b> <button type="button" class="delete" id="{{newcheep._id}}"user="{{newcheep.username}}" uid="{{newcheep.userid}}">Delete</button>
-  <p>{{newcheep.words}}</p>
-</div>
