+++ views/home.handlebars
@@ -1,32 +1,10 @@
-  {{> header}}
+{{> header}}
 
-  <div id="mainpage">
-    <div id= "feed">
-      <div class = "cheep newcheep">
-        <form id="newcheep">
-          <textarea name="in_cheep" id="in_cheep" placeholder="Say something cheep, {{currentuser.name}}..." uid="{{currentuser._id}}" autofocus></textarea>
-          <div id="buttondiv"><button class="send" type="submit" id="cheepbutton">Cheep!</button></div>
-        </form>
-      </div>
-
-      {{#each cheeps}}
-      <div class="cheep" user="{{this.username}}" id="{{this._id}}" uid="{{this.userid}}">
-        <b>{{this.username}} said... </b> <button type="button" class="delete" id="{{this._id}}"user="{{this.username}}" uid="{{this.userid}}">Delete</button>
-        <p>{{this.words}}</p>
-      </div>
-      {{/each}}
+<div class="container">
+    <div class="row">
+        <div class="col-md-12">
+          <div class="well">
+          </div>
+        </div>
     </div>
-
-    <div id="users">
-      {{#each users}}
-      <div class="user" user="{{this.name}}" uid="{{this._id}}">
-        <div class="usernamedisp" user="{{this.name}}" uid="{{this._id}}"><b>{{this.name}}</b></div>
-      </div>
-      {{/each}}
-      <div class="user clhl" user="{{this.name}}">
-        <div id="clearuserhl" user="clearhighlight"><b>Clear Highlight</b></div>
-      </div>
-    </div>
-
-  </div>
-
+</div>
