+++ views/layouts/main.handlebars
@@ -2,8 +2,13 @@
 <html lang="en">
 <head>
   <meta charset="utf-8"/>
-  <title>Cheepr</title>
-  <link rel="icon" href="images/favicon.ico" type="image/x-icon" />
+  <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no">
+  <title>giftr</title>
+  <link rel="icon" type="image/x-icon" href="images/favicon.ico">
+  <link rel="icon" type="image/png" sizes="32x32" href="images/favicon-32x32.png">
+  <link rel="icon" type="image/png" sizes="96x96" href="images/favicon-96x96.png">
+  <link rel="icon" type="image/png" sizes="16x16" href="images/favicon-16x16.png">
+
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.2/css/bootstrap.min.css">
   <link href="/stylesheets/ripples.min.css" rel="stylesheet">
   <link href="/stylesheets/material-wfont.min.css" rel="stylesheet">
