+++ views/login.handlebars
@@ -1,8 +1,25 @@
 {{> header}}
-<div id="mainpage">
-  <div class="auth">
-    <h1> Log in to giftr </h1>
-    <a href="/auth/facebook" class="btn btn-lg btn-primary btn-raised">Log in!</a>
-    <div><a href="/auth/facebook">Sign in with Facebook</a></div>
-  </div>
+<!-- <div id="mainpage">
+    <div class="auth">
+        <h1> Log in to giftr </h1>
+        <a href="/auth/facebook" class="btn btn-lg btn-primary btn-raised">Log in!</a>
+        <div><a href="/auth/facebook">Sign in with Facebook</a>
+        </div>
+    </div>
+</div> -->
+<div class="container">
+    <div class="row">
+        <div class="col-md-4"></div>
+        <div class="col-md-4">
+            <div class="well">
+                <h1> Log in to giftr </h1>
+                <div class="text-center">
+                    <a href="/auth/facebook" class="btn btn-lg btn-primary btn-raised">Log in!</a>
+                </div>
+                <div><a href="/auth/facebook">Sign in with Facebook</a>
+                </div>
+            </div>
+        </div>
+        <div class="col-md-4"></div>
+    </div>
 </div>
