+++ views/partials/header.handlebars
@@ -1,4 +1,4 @@
-<div class="navbar navbar-default navbar-fixed-top shadow-z-2">
+<div class="navbar navbar-default shadow-z-2">
     <div class="container-fluid">
         <div class="navbar-header">
             <button type="button" class="navbar-toggle" data-toggle="collapse" data-target=".navbar-responsive-collapse">
